+++ app/controllers/vendor_api/test_data_controller.rb
@@ -24,6 +24,8 @@ module VendorApi
       render json: { data: { ids: application_choices.map { |ac| ac.id.to_s } } }
     rescue TestApplications::NotEnoughCoursesError => e
       render json: { errors: [{ error: 'ParameterInvalid', message: e }] }, status: :unprocessable_entity
+    rescue TestApplications::ZeroCoursesPerApplicationError => e
+      render json: { errors: [{ error: 'ParameterInvalid', message: e }] }, status: :unprocessable_entity
     end
 
     def clear!